Moving helper functions to a separate file
