+++ common/app/navigation/SectionLinks.scala
@@ -1,0 +1,89 @@
+package common
+
+import NavLinks._
+import NewNavigation._
+
+object SectionLinks {
+
+  val sectionLinks = List (
+    SectionsLink ("uk", headlines, MostPopular),
+    SectionsLink ("us", headlines, MostPopular),
+    SectionsLink ("au", headlines, MostPopular),
+    SectionsLink ("international", headlines, MostPopular),
+    SectionsLink ("uk-news", ukNews, News),
+    SectionsLink ("world", world, News),
+    SectionsLink ("world/europe-news", europe, News),
+    SectionsLink ("politics", politics, News),
+    SectionsLink ("environment", environment, News),
+    SectionsLink ("business", business, News),
+    SectionsLink ("technology", tech, News),
+    SectionsLink ("science", science, News),
+    SectionsLink ("money", money, News),
+    SectionsLink ("australia-news", australiaNews, News),
+    SectionsLink ("media", media, News),
+    SectionsLink ("us-news", usNews, News),
+    SectionsLink ("cities", cities, News),
+    SectionsLink ("inequality", inequality, News),
+    SectionsLink ("global-development", globalDevelopment, News),
+    SectionsLink ("sustainable-business", sustainableBusiness, News),
+    SectionsLink ("law", law, News),
+    SectionsLink ("technology/games", games, News),
+    SectionsLink ("us-news/us-politics", usPolitics, News),
+    SectionsLink ("australia-news/australian-politics", auPolitics, News),
+    SectionsLink ("australia-news/australian-immigration-and-asylum", auImmigration, News),
+    SectionsLink ("australia-news/indigenous-australians", indigenousAustralia, News),
+
+    SectionsLink ("commentisfree", opinion, Opinion),
+    SectionsLink ("cartoons/archive", cartoons, Opinion),
+    SectionsLink ("type/cartoon", cartoons, Opinion),
+    SectionsLink ("au/index/contributors", auColumnists, Opinion),
+    SectionsLink ("index/contributors", columnists, Opinion),
+    SectionsLink ("commentisfree/series/comment-is-free-weekly", inMyOpinion, Opinion),
+    SectionsLink ("profile/editorial", theGuardianView, Opinion),
+
+    SectionsLink ("sport", sport, Sport),
+    SectionsLink ("football", football, Sport),
+    SectionsLink ("sport/rugby-union", rugbyUnion, Sport),
+    SectionsLink ("sport/cricket", cricket, Sport),
+    SectionsLink ("sport/tennis", tennis, Sport),
+    SectionsLink ("sport/golf", golf, Sport),
+    SectionsLink ("sport/us-sport", usSports, Sport),
+    SectionsLink ("sport/horse-racing", racing, Sport),
+    SectionsLink ("sport/rugbyleague", rugbyLeague, Sport),
+    SectionsLink ("sport/boxing", boxing, Sport),
+    SectionsLink ("sport/formulaone", formulaOne, Sport),
+    SectionsLink ("sport/nfl", NFL, Sport),
+    SectionsLink ("sport/mlb", MLB, Sport),
+    SectionsLink ("football/mls", MLS, Sport),
+    SectionsLink ("sport/nba", NBA, Sport),
+    SectionsLink ("sport/nhl", NHL, Sport),
+    SectionsLink ("sport/afl", AFL, Sport),
+    SectionsLink ("football/a-league", aLeague, Sport),
+    SectionsLink ("sport/nrl", NRL, Sport),
+    SectionsLink ("sport/australia-sport", australiaSport, Sport),
+
+    SectionsLink ("culture", culture, Arts),
+    SectionsLink ("film", film, Arts),
+    SectionsLink ("tv-and-radio", tvAndRadio, Arts),
+    SectionsLink ("music", music, Arts),
+    SectionsLink ("books", books, Arts),
+    SectionsLink ("artanddesign", artAndDesign, Arts),
+    SectionsLink ("stage", stage, Arts),
+    SectionsLink ("music/classicalmusicandopera", classical, Arts),
+
+    SectionsLink ("lifeandstyle", lifestyle, Life),
+    SectionsLink ("fashion", fashion, Life),
+    SectionsLink ("travel", travel, Life),
+    SectionsLink ("society", society, Life),
+    SectionsLink ("lifeandstyle/food-and-drink", food, Life),
+    SectionsLink ("tone/recipes", recipes, Life),
+    SectionsLink ("lifeandstyle/women", women, Life),
+    SectionsLink ("lifeandstyle/health-and-wellbeing", health, Life),
+    SectionsLink ("lifeandstyle/family", family, Life),
+    SectionsLink ("lifeandstyle/love-and-sex", loveAndSex, Life),
+    SectionsLink ("au/lifeandstyle/fashion", fashionAu, Life),
+    SectionsLink ("au/lifeandstyle/food-and-drink", foodAu, Life),
+    SectionsLink ("au/lifeandstyle/relationships", relationshipsAu, Life),
+    SectionsLink ("au/lifeandstyle/health-and-wellbeing", healthAu, Life)
+  )
+}
