+++ common/app/views/fragments/newHeader.scala.html
@@ -1,7 +1,7 @@
 @(page: model.Page)(implicit request: RequestHeader)
 
 @import conf.Configuration
-@import common.{LinkTo, NewNavigation, Edition}
+@import common.{LinkTo, NewNavigation, Edition, NavigationHelpers}
 @import views.support.RenderClasses
 
 @jobsUrl(edition: String) = {
@@ -25,9 +25,9 @@
             "new-header--slim" -> page.metadata.hasSlimHeader
         ), "new-header")"
         role="banner">
-    @defining(NewNavigation.SubSectionLinks.getSectionOrTagId(page)) { id =>
+    @defining(NavigationHelpers.getSectionOrPageId(page)) { id =>
         @defining(
-            NewNavigation.SectionLinks.getPillarName(id)
+            NavigationHelpers.getPillarName(id)
         ) { case (currentTopLevelSection) =>
             <nav class="new-header__inner gs-container"
                  data-component="nav2"
